backend/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
sqlalchemy
firebase-admin
